Move Event methods in ContestedDataBuilder to EventApi
